combine FIleManager and fileProcesor
